move group queries into their respective component folders
