+++ src/groups/GroupContentItemConnected/getGroup.js
