+++ src/groups/GroupListConnected/getCurrentUserGroups.js
